+++ src/helpers/resolve_projects.rs
@@ -1,17 +1,17 @@
 // Copyright (C) 2026 Lordseriouspig
-// 
+//
 // This file is part of flavorcli.
-// 
+//
 // flavorcli is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
 // the Free Software Foundation, either version 3 of the License, or
 // (at your option) any later version.
-// 
+//
 // flavorcli is distributed in the hope that it will be useful,
 // but WITHOUT ANY WARRANTY; without even the implied warranty of
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
-// 
+//
 // You should have received a copy of the GNU General Public License
 // along with flavorcli.  If not, see <https://www.gnu.org/licenses/>.
 
@@ -20,7 +20,7 @@ use anyhow::Result;
 use indicatif::{ProgressBar, ProgressStyle};
 use log::{debug, info};
 
-pub async fn resolve_projects(project_ids: &[u64]) -> Result<Vec<Project>> {
+pub async fn resolve_projects(project_ids: &[u32]) -> Result<Vec<Project>> {
     let auth: AuthData = get_key()?;
     let spinner = ProgressBar::new_spinner();
     let mut resolved_projects = Vec::new();
@@ -32,7 +32,10 @@ pub async fn resolve_projects(project_ids: &[u64]) -> Result<Vec<Project>> {
     spinner.enable_steady_tick(std::time::Duration::from_millis(80));
     let client = reqwest::Client::new();
     for project in project_ids {
-        let url = format!("https://flavortown.hackclub.com/api/v1/projects/{}", project);
+        let url = format!(
+            "https://flavortown.hackclub.com/api/v1/projects/{}",
+            project
+        );
         debug!("Sending GET request to {}", url);
         let res = client
             .get(&url)
